Decouple client from underlying implementation
Fix #12

Update the SecretClient to better abstract details regarding the source
of the secrets, like Vault or a standard file.

Reconstruct the SecretClient so that the concerns are layered to allow
for easier testing. The SecretClient is responsible for business logic,
like how to handle a non-existent key during retrieval. The
SecretStoreManager interface abstracts the details of how the keys are
stored(HTTP, File, Database, etc) and how they are managed.

Add tests for the SecretClient business logic, and the
httpSecretStoreManager

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>